move delete button from workspace overview page to update page
