+++ src/plugins/workspace/public/components/workspace_overview.tsx
@@ -10,7 +10,6 @@ import { of } from 'rxjs';
 import { i18n } from '@osd/i18n';
 import { ApplicationStart } from '../../../../core/public';
 import { useOpenSearchDashboards } from '../../../opensearch_dashboards_react/public';
-import { DeleteWorkspaceModal } from './delete_workspace_modal';
 import { PATHS } from '../../common/constants';
 import { WORKSPACE_APP_ID, WORKSPACE_ID_IN_SESSION_STORAGE } from '../../common/constants';
 
@@ -22,43 +21,6 @@ export const WorkspaceOverview = () => {
   const currentWorkspace = useObservable(
     workspaces ? workspaces.client.currentWorkspace$ : of(null)
   );
-
-  const workspaceId = currentWorkspace?.id;
-  const workspaceName = currentWorkspace?.name;
-  const [deleteWorkspaceModalVisible, setDeleteWorkspaceModalVisible] = useState(false);
-
-  const deleteWorkspace = async () => {
-    if (workspaceId) {
-      let result;
-      try {
-        result = await workspaces?.client.delete(workspaceId);
-      } catch (error) {
-        notifications?.toasts.addDanger({
-          title: i18n.translate('workspace.delete.failed', {
-            defaultMessage: 'Failed to delete workspace',
-          }),
-          text: error instanceof Error ? error.message : JSON.stringify(error),
-        });
-        return setDeleteWorkspaceModalVisible(false);
-      }
-      if (result?.success) {
-        notifications?.toasts.addSuccess({
-          title: i18n.translate('workspace.delete.success', {
-            defaultMessage: 'Delete workspace successfully',
-          }),
-        });
-      } else {
-        notifications?.toasts.addDanger({
-          title: i18n.translate('workspace.delete.failed', {
-            defaultMessage: 'Failed to delete workspace',
-          }),
-          text: result?.error,
-        });
-      }
-    }
-    setDeleteWorkspaceModalVisible(false);
-    await application.navigateToApp('home');
-  };
 
   const onUpdateWorkspaceClick = () => {
     if (!currentWorkspace || !currentWorkspace.id) {
@@ -78,23 +40,9 @@ export const WorkspaceOverview = () => {
     <>
       <EuiPageHeader
         pageTitle="Overview"
-        rightSideItems={[
-          <EuiButton color="danger" onClick={() => setDeleteWorkspaceModalVisible(true)}>
-            Delete
-          </EuiButton>,
-          <EuiButton>Update</EuiButton>,
-          <EuiButton color="danger">Delete</EuiButton>,
-          <EuiButton onClick={onUpdateWorkspaceClick}>Update</EuiButton>,
-        ]}
+        rightSideItems={[<EuiButton onClick={onUpdateWorkspaceClick}>Update</EuiButton>]}
       />
       <EuiPanel>
-        {deleteWorkspaceModalVisible && (
-          <DeleteWorkspaceModal
-            onConfirm={deleteWorkspace}
-            onClose={() => setDeleteWorkspaceModalVisible(false)}
-            selectedItems={workspaceName ? [workspaceName] : []}
-          />
-        )}
         <EuiTitle size="m">
           <h3>Workspace</h3>
         </EuiTitle>
